+++ frontend/test/pytest/test_conditionals.py
@@ -24,12 +24,7 @@ class TestCondToJaxpr:
         expected = """{ lambda ; a:i64[]. let
     b:bool[] = eq a 5
     c:i64[] = qcond[
-      false_jaxpr={ lambda ; d_:i64[] e:i64[]. let
-          f:i64[] = integer_pow[y=3] e
-        in (f,) }
-      true_jaxpr={ lambda ; g:i64[] h_:i64[]. let
-          i:i64[] = integer_pow[y=2] g
-        in (i,) }
+      branch_jaxprs=[{ lambda ; a:i64[] b_:i64[]. let c:i64[] = integer_pow[y=2] a in (c,) }, { lambda ; a_:i64[] b:i64[]. let c:i64[] = integer_pow[y=3] b in (c,) }]
     ] b a a
   in (c,) }"""
 
@@ -72,6 +67,87 @@ class TestCond:
         assert circuit(5) == 25
         assert circuit(6) == 36
 
+    def test_cond_one_else_if(self):
+        """Test a cond with one else_if branch"""
+
+        @qjit
+        @qml.qnode(qml.device("lightning.qubit", wires=1))
+        def circuit(x):
+            @cond(x > 2.7)
+            def cond_fn():
+                return x * 4
+
+            @cond_fn.else_if(x > 1.4)
+            def cond_elif():
+                return x * 2
+
+            @cond_fn.otherwise
+            def cond_else():
+                return x
+
+            return cond_fn()
+
+        assert circuit(4) == 16
+        assert circuit(2) == 4
+        assert circuit(1) == 1
+
+    def test_cond_many_else_if(self):
+        """Test a cond with multiple else_if branches"""
+
+        @qjit
+        @qml.qnode(qml.device("lightning.qubit", wires=1))
+        def circuit(x):
+            @cond(x > 4.8)
+            def cond_fn():
+                return x * 8
+
+            @cond_fn.else_if(x > 2.7)
+            def cond_elif():
+                return x * 4
+
+            @cond_fn.else_if(x > 1.4)
+            def cond_elif2():
+                return x * 2
+
+            @cond_fn.otherwise
+            def cond_else():
+                return x
+
+            return cond_fn()
+
+        assert circuit(5) == 40
+        assert circuit(3) == 12
+        assert circuit(2) == 4
+        assert circuit(-3) == -3
+
+    def test_cond_else_if_classical(self):
+        """Test a cond with multiple else_if branches using the classical compilation path."""
+
+        @qjit
+        def circuit(x):
+            @cond(x > 4.8)
+            def cond_fn():
+                return x * 16
+
+            @cond_fn.else_if(x > 2.7)
+            def cond_elif():
+                return x * 8
+
+            @cond_fn.else_if(x > 1.4)
+            def cond_elif2():
+                return x * 4
+
+            @cond_fn.otherwise
+            def cond_else():
+                return x
+
+            return cond_fn()
+
+        assert circuit(5) == 80
+        assert circuit(3) == 24
+        assert circuit(2) == 8
+        assert circuit(-3) == -3
+
     def test_qubit_manipulation_cond(self):
         @qjit()
         @qml.qnode(qml.device("lightning.qubit", wires=1))
@@ -88,6 +164,11 @@ class TestCond:
         assert circuit(6) == True
 
     def test_branch_return_mismatch(self):
+        """
+        Test that an exception is raised when the true branch returns a value without an else
+        branch
+        """
+
         def circuit():
             @cond(True)
             def cond_fn():
@@ -95,7 +176,51 @@ class TestCond:
 
             return cond_fn()
 
-        with pytest.raises(TypeError, match="Conditional branches require the same return type"):
+        with pytest.raises(
+            TypeError, match="Conditional branches all require the same return type"
+        ):
+            qjit(qml.qnode(qml.device("lightning.qubit", wires=1))(circuit))
+
+    def test_branch_multi_return_mismatch(self):
+        """Test that an exception is raised when the return types of all branches do not match"""
+
+        def circuit():
+            @cond(True)
+            def cond_fn():
+                return measure(wires=0)
+
+            @cond_fn.else_if(False)
+            def cond_elif():
+                return 0
+
+            @cond_fn.otherwise
+            def cond_else():
+                return measure(wires=0)
+
+            return cond_fn()
+
+        with pytest.raises(
+            TypeError, match="Conditional branches all require the same return type"
+        ):
+            qjit(qml.qnode(qml.device("lightning.qubit", wires=1))(circuit))
+
+    def test_branch_with_arg(self):
+        """Test that an exception is raised when an 'else if' branch function contains an arg"""
+
+        def circuit(pred: bool):
+            @cond(pred)
+            def cond_fn():
+                qml.PauliX(0)
+
+            @cond_fn.else_if(pred)
+            def cond_elif(x):
+                qml.PauliX(x)
+
+            return measure(wires=0)
+
+        with pytest.raises(
+            TypeError, match="Conditional 'else if' function is not allowed to have any arguments"
+        ):
             qjit(qml.qnode(qml.device("lightning.qubit", wires=1))(circuit))
 
     def test_identical_branch_names(self):